# api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Auction System API
  version: 1.0.0
  description: Production-grade auction platform — all endpoints
servers:
  - url: http://localhost:8080
    description: Local (API Gateway)
...
paths:
  /api/v1/auth/register:
    post: {tags: [Auth], summary: Register, requestBody: {required: true, content: {application/json: {schema: {type: object, required: [username,email,password,first_name,last_name], properties: {username: {type: string}, email: {type: string, format: email}, password: {type: string, minLength: 8}, first_name: {type: string}, last_name: {type: string}, role: {type: string, enum: [bidder,seller]}}}}}}, responses: {'201': {description: Registered}}}
  /api/v1/auth/login:
    post: {tags: [Auth], summary: Login, requestBody: {required: true, content: {application/json: {schema: {type: object, required: [email,password], properties: {email: {type: string}, password: {type: string}}}}}}, responses: {'200': {description: Token pair}}}
  /api/v1/auth/refresh:
    post: {tags: [Auth], summary: Refresh token, requestBody: {content: {application/json: {schema: {type: object, properties: {refresh_token: {type: string}}}}}}, responses: {'200': {description: New tokens}}}
  /api/v1/auth/logout:
    post: {tags: [Auth], summary: Logout, security: [{BearerAuth: []}], responses: {'200': {description: Logged out}}}
...
  /api/v1/users/me:
    get: {tags: [Users], summary: Get my profile, security: [{BearerAuth: []}], responses: {'200': {description: Profile}}}
    put: {tags: [Users], summary: Update profile, security: [{BearerAuth: []}], responses: {'200': {description: Updated}}}
    delete: {tags: [Users], summary: Delete account, security: [{BearerAuth: []}], responses: {'200': {description: Deleted}}}
...
  /api/v1/users/{id}:
    get: {tags: [Users], summary: Public profile, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Public profile}}}
  /api/v1/auctions:
    get: {tags: [Auctions], summary: List auctions, parameters: [{name: state, in: query, schema: {type: string, enum: [active,scheduled,ended]}}, {name: page, in: query, schema: {type: integer}}, {name: page_size, in: query, schema: {type: integer}}], responses: {'200': {description: List}}}
    post: {tags: [Auctions], summary: Create auction (seller), security: [{BearerAuth: []}], requestBody: {required: true, content: {application/json: {schema: {type: object, required: [title,description,starting_price,start_time,end_time], properties: {title: {type: string}, description: {type: string}, starting_price: {type: number}, reserve_price: {type: number}, start_time: {type: string, format: date-time}, end_time: {type: string, format: date-time}}}}}}, responses: {'201': {description: Created}}}
  /api/v1/auctions/{id}:
    get: {tags: [Auctions], summary: Get auction, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Auction}}}
    put: {tags: [Auctions], summary: Update auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Updated}}}
    delete: {tags: [Auctions], summary: Delete auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Deleted}}}
...
  /api/v1/seller/auctions:
    get: {tags: [Auctions], summary: My auctions (seller), security: [{BearerAuth: []}], responses: {'200': {description: Seller auctions}}}
...
  /api/v1/bids:
    post: {tags: [Bids], summary: Place bid, security: [{BearerAuth: []}], requestBody: {required: true, content: {application/json: {schema: {type: object, required: [auction_id,amount], properties: {auction_id: {type: string}, amount: {type: number}}}}}}, responses: {'201': {description: Bid placed}, '400': {description: Bid too low}, '429': {description: Auction busy}}}
  /api/v1/bids/me:
    get: {tags: [Bids], summary: My bids, security: [{BearerAuth: []}], responses: {'200': {description: Bid history}}}
  /api/v1/auctions/{id}/bids:
    get: {tags: [Bids], summary: Auction bids, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Bids}}}
...
  /api/v1/ws/{auction_id}:
    get: {tags: [Bids], summary: WebSocket real-time bids, security: [{BearerAuth: []}], parameters: [{name: auction_id, in: path, required: true, schema: {type: string}}], responses: {'101': {description: WebSocket upgrade}}}
  /api/v1/wallet:
    get: {tags: [Wallet], summary: Get balance, security: [{BearerAuth: []}], responses: {'200': {description: Wallet}}}
  /api/v1/wallet/deposit:
    post: {tags: [Wallet], summary: Deposit, security: [{BearerAuth: []}], requestBody: {content: {application/json: {schema: {type: object, properties: {amount: {type: number}}}}}}, responses: {'200': {description: Deposited}}}
...
  /api/v1/wallet/history:
    get: {tags: [Wallet], summary: Transaction history, security: [{BearerAuth: []}], responses: {'200': {description: Transactions}}}
...
  /api/v1/search:
    get: {tags: [Search], summary: Search auctions, parameters: [{name: q, in: query, schema: {type: string}}, {name: category, in: query, schema: {type: string}}, {name: min_price, in: query, schema: {type: number}}, {name: max_price, in: query, schema: {type: number}}, {name: sort_by, in: query, schema: {type: string, enum: [price_asc,price_desc,newest,ending_soon,most_bids]}}], responses: {'200': {description: Results}}}
...
  /api/v1/analytics/admin/dashboard:
    get: {tags: [Analytics], summary: Admin KPIs, security: [{BearerAuth: []}], responses: {'200': {description: Dashboard}}}
  /api/v1/analytics/revenue:
    get: {tags: [Analytics], summary: Revenue metrics, security: [{BearerAuth: []}], parameters: [{name: from, in: query, schema: {type: string, format: date}}, {name: to, in: query, schema: {type: string, format: date}}], responses: {'200': {description: Revenue}}}
  /api/v1/analytics/seller/{id}:
    get: {tags: [Analytics], summary: Seller analytics, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Seller stats}}}
...
  /api/v1/admin/users:
    get: {tags: [Admin], summary: List all users, security: [{BearerAuth: []}], responses: {'200': {description: Users}}}
  /api/v1/admin/users/{id}/ban:
    put: {tags: [Admin], summary: Ban user, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Banned}}}
  /api/v1/admin/auctions:
    get: {tags: [Admin], summary: All auctions (admin), security: [{BearerAuth: []}], responses: {'200': {description: Auctions}}}
  /api/v1/admin/auctions/{id}:
    delete: {tags: [Admin], summary: Force delete auction, security: [{BearerAuth: []}], parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Deleted}}}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT